of: configfs: Use of_overlay_fdt_apply API call
The published API to the dynamic overlay application mechanism
now takes a Flattened Device Tree blob as input so that it can
manage the lifetime of the unflattened tree. So this patch
replaces the new API call of_overlay_fdt_apply with unflatten
device-tree handling logic. It uses Call of_overlay_remove()
to remove and clean-up the overlay changeset previously created
via the call of_overlay_fdt_apply().

Signed-off-by: Nava kishore Manne <nava.manne@xilinx.com>
State: not-upstreamable
Signed-off-by: Isaac True <isaac.true@canonical.com>