ARM: dts: at91: sam9x60: Disable SHA IP
There are some bugs on the SHA IP on sam9x60 which are not yet
handled. On linux4sam-2020.10 CONFIG_CRYPTO_DEV_ATMEL_SHA was
disabled, and the driver never probed. Recently it was enabled.
Since the hw implementation has a higher priority than the software
SHA implementation, the enablement of the SHA driver can introduce
regressions. Disable the SHA IP until we'll fix all the reported bugs.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>
Acked-by: Cristian Birsan <cristian.birsan@microchip.com>
[claudiu.beznea: adapt to 5.15]
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Signed-off-by: Isaac True <isaac.true@canonical.com>