clk: at91: sama7g5: do not use AUDIOPLL as changeable parent
AUDIOPLL can be a parrent to many GCLKs. This causes a race on AUDIOPLL,
since the first audio IP that uses it will set the AUDIOPLL to the lowest
rate that can be served and would be acceptable by the requesting audio
IP. This rate might not be good for another audio IP that wants to use the
AUDIOPLL as its parent, so its clk_set_rate() will return an error.
Because of this, we remove the ability to change the rate of the AUDIOPLL.
AUDIOPLL can be set using assigned-clocks to a value that is big enough to
cover all the audio IPs needs and enable the use of GCLKs divisors.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@microchip.com>
Reviewed-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Signed-off-by: Isaac True <isaac.true@canonical.com>