media: atmel: atmel-isc-base: implement mbus_code support in enumfmt
If enumfmt is called with an mbus_code, the enumfmt handler should only
return the formats that are supported for this mbus_code.
To make it more easy to understand the formats, changed the report order
to report first the native formats, and after that the formats that the ISC
can convert to.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>
Signed-off-by: Isaac True <isaac.true@canonical.com>