ARM: at91: pm: fix self-refresh for sama7g5
It has been discovered that on some parts, from time to time,
self-refresh procedure doesn't work as expected. Debugging it
it has been proved that reverting DLL bypass and power down of
the IOs fix the issue. Reverting only one or the other had no
effect (issue still present). Thus, for the moment keep only
the DX DLLs bypassing and avoid AC DLL bypass + IO disable.
It may happen that AC DLL bypass to produce glitches in RAM
controllers that might produce the breakage that has been
discovered. With the code in current patch there is an extra
current consumption of 6mA when switching to any ULP or standby
modes but the self-refresh procedure still works. Thus until
further investigation keep it like this.

Fixes: f0bbf17958e8 ("ARM: at91: pm: add self-refresh support for sama7g5")
Suggested-by: Frederic Schumacher <frederic.schumacher@microchip.com>
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Tested-by: Cristian Birsan <cristian.birsan@microchip.com>
Signed-off-by: Isaac True <isaac.true@canonical.com>