spi: microchip-core: fix chip select behaviour
The existing code does not allow the spi framework to change the chip
select as it likes & instead arbitrarily turns the chip select on and
off in {un,}prepare_message().

Allow the framework to manipulate the chip select by creating a
set_cs() function & removing control from {un,}prepare_message().
Direct mode (where the hardware *doesn't* control the chip select) now
has to be enabled during initialisation not during prepare_message()
so move it there.

As a side effect of this change, the {en,dis}able}_direct_mode(),
unprepare_message()  & {de,}activate_cs() functions are now useless
and so can be removed.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Isaac True <isaac.true@canonical.com>