media: microchip: mchp-dscmi: stop reporting invalid restart failures
When capture is started for the first time, the DMA buffer queue may not
be populated yet. If this happens, an error is logged to the console:

mchp-dscmi 40000000.video-pipeline: failed to mchp restart capture -22

Change mchp_dscmi_restart_capture() to return zero instead of -EINVAL
if the mchp_dscmi->buf_list queue is empty. The error will still be
tracked as a dropped "packet" for debug purposes.

Signed-off-by: Shravan Chippa <shravan.chippa@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Isaac True <isaac.true@canonical.com>